Aligning the angle syntax with the dollar syntax
I also harmonized some whitespaces.

This is the best TM grammar I could find. I'd love to copy this grammar to the [Theia project](http://www.theia-ide.org/). Theia is an open-source, cloud-based IDE framework that could stand some basic StringTemplate support.  The problem is, you haven't put a license on this. So if you want it to be adopted, please put on some permissive open-source license (MIT would do).